fix(CartSummary/singlecartItem)fixed cart summary page with add to cart function

+++ server/db/models/product.js
@@ -18,7 +18,7 @@ const Product = db.define('product', {
       'https://images.clipartlogo.com/files/istock/previews/4066/40662726-bakery-shop.jpg'
   },
   currentPrice: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(12, 2),
     allowNull: false
   },
   quantity: {
